pyproject.toml
[project]
name = "trackit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Gustav R. Jansen", email = "gustav.r.jansen@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.0",
    "sqlalchemy>=2.0.0",
    "python-dateutil>=2.8.0",
]

[project.scripts]
trackit = "trackit:main"

[build-system]
requires = ["uv_build>=0.9.15,<0.10.0"]
build-backend = "uv_build"
